+++ frontend/app/components/GrowthMaturityDeclineCard.vue
@@ -5,11 +5,11 @@
     <div class="row">
 
       <div class="col col-6">
-        <area-line-chart source="issues"
+        <line-chart source="issues"
                     title="Issues / Week"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/open-issues.md"
                     cite-text="Issues Open">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
@@ -21,88 +21,88 @@
       </div>
 
       <!-- <div class="col col-6">
-        <area-line-chart source="issuesResponseTime"
+        <line-chart source="issuesResponseTime"
                     title="Issue Response Time"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/issue-response-time.md"
                     cite-text="Issue Response Time">
-        </area-line-chart>
+        </line-chart>
       </div> -->
 
       <div class="col col-6">
-        <area-line-chart source="commits"
+        <line-chart source="commits"
                     title="Commits / Week"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/commits.md"
                     cite-text="Commits">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="forks"
+        <line-chart source="forks"
                     title="Forks / Week"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/forks.md"
                     cite-text="Forks">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="pulls"
+        <line-chart source="pulls"
                     title="Pulls / Week"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/pull-requests-open.md"
                     cite-text="Open Pull Requests">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="pullReqComments"
+        <line-chart source="pullReqComments"
                     title="Pull Request Comments / Week "
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/pull-request-comments.md"
                     cite-text="Pull Request Comments">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <!-- <div class="col col-6">
-        <area-line-chart source="codeReviewIteration"
+        <line-chart source="codeReviewIteration"
                       title="Number of Code Review Iterations"
                       size="total"
                       cite-url="https://github.com/chaoss/metrics/blob/master/activity-metrics/code-review-iteration.md"
                       cite-text="Code Review Iterations">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="maintainerResponseTime"
+        <line-chart source="maintainerResponseTime"
                       title="Time to First Maintainer Response to Merge Request"
                       size="total"
                       cite-url="https://github.com/chaoss/metrics/blob/master/activity-metrics/maintainer-response-to-merge-request-duration.md"
                       cite-text="Time to First Maintainer Response to Merge Request">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="contributionAcceptance"
+        <line-chart source="contributionAcceptance"
                       title="Contribution Acceptance"
                       size="total"
                       cite-url="https://github.com/chaoss/metrics/blob/master/activity-metrics/contribution-acceptance.md"
                       cite-text="Contribution Acceptance">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="newContributingGithubOrganizations"
+        <line-chart source="newContributingGithubOrganizations"
                       title="New Contributing Github Organizations"
                       size="total"
                       cite-url="https://github.com/chaoss/metrics/blob/master/activity-metrics/new-contributing-organizations.md"
                       cite-text="New Contributing Organizations">
-        </area-line-chart>
+        </line-chart>
       </div> -->
 
       <!-- <div class="col col-6">
-        <area-line-chart source="codeReviews"
+        <line-chart source="codeReviews"
                       title="Code Reviews"
                       size="total"
                       cite-url="https://github.com/chaoss/metrics/blob/master/activity-metrics/code-reviews.md"
                       cite-text="Code Reviews">
-        </area-line-chart>
+        </line-chart>
       </div> -->
 
       <div class="col col-12">
@@ -115,71 +115,71 @@
       </div>
 <!--
       <div class="col col-6">
-        <area-line-chart source="communityEngagement:issues_open"
+        <line-chart source="communityEngagement:issues_open"
                     title="Open Issues"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/open-issues.md"
                     cite-text="Open Issues"
                     disable-rolling-average=1>
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="communityEngagement:issues_closed_total"
+        <line-chart source="communityEngagement:issues_closed_total"
                     title="Closed Issues"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/closed-issues.md"
                     cite-text="Closed Issues"
                     disable-rolling-average=1>
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="issueComments"
+        <line-chart source="issueComments"
                     title="Issue Comments / Week "
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/people-opening-issues.md"
                     cite-text="Issue Comments">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="commitComments"
+        <line-chart source="commitComments"
                     title="Commit Comments / Week "
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/code-reviews.md"
                     cite-text="Code Reviews (naive implementation)">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="pullReqComments"
+        <line-chart source="pullReqComments"
                     title="Pull Request Comments / Week "
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/pull-request-comments.md"
                     cite-text="Pull Request Comments">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="pullsAcceptanceRate"
+        <line-chart source="pullsAcceptanceRate"
                     title="Pull Request Acceptance Rate"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/pull-requests-made-closed.md"
                     cite-text="Pull Requets Made / Closed">
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="area-linesChanged:total_area-lines"
+        <line-chart source="linesChanged:total_lines"
                     title="Size of Code Base"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/size-of-code-base.md"
                     cite-text="Size of Code Base"
                     disableRollingAverage=1>
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-6">
-        <area-line-chart source="totalCommitters"
+        <line-chart source="totalCommitters"
                     title="Total Committers"
                     cite-url="https://github.com/OSSHealth/wg-gmd/blob/master/activity-metrics/contributors.md"
                     cite-text="Contributors"
                     disableRollingAverage=1>
-        </area-line-chart>
+        </line-chart>
       </div>
 
       <div class="col col-12">
