+++ frontend/app/components/charts/LineChart.vue
@@ -1,33 +1,33 @@
 <template>
   <div ref="holder">
-    <div ref="chart" class="linechart">
-      <div ref="legend" class="legend hideme invis"></div>
-      <div ref="chartholder"></div>
-      <span ref="chartStatus" class="showme" v-html="chart"></span>
+    <div class="hidefirst invis">
+      <vega-lite :spec="spec" :data="values"></vega-lite>
+      <p> {{ chart }} </p>
     </div>
-    <div class="row below-chart hideme invis">
-      <div class="col col-6"><cite class="metric">Metric: <a v-bind:href="citeUrl" target="_blank">{{ citeText }}</a></cite></div>
+
+    <div class="row below-chart">
+      <div class="col col-5"><cite class="metric">Metric: <a v-bind:href="citeUrl" target="_blank">{{ citeText }}</a></cite></div>
       <div class="col col-6"><button class="button download graph-download" v-on:click="downloadSVG">&#11015; SVG</button><button class="button graph-download download" v-on:click="downloadPNG">&#11015; PNG</button><a class="button graph-download download" ref="downloadJSON" role="button">&#11015; JSON</a></div>
+      <!-- <div class="form-item form-checkboxes">
+        <label class="checkbox"><input name="hidearea" value="each" type="checkbox" v-model="ar">Individual area<sup class="warn"></sup></label><br>
+      </div> -->
     </div>
+
   </div>
 </template>
 
 
 <script>
+import { mapState } from 'vuex'
 import AugurStats from 'AugurStats'
 
+
 export default {
-  props: ['source', 'citeUrl', 'citeText', 'title', 'percentage',
-    'comparedTo', 'disableRollingAverage', 'alwaysByDate', 'innerKey'],
-  data () {
+  props: ['source', 'citeUrl', 'citeText', 'title', 'disableRollingAverage', 'alwaysByDate'],
+  data() {
     return {
-      mgConfig: {
-        time_series: true,
-        full_width: true,
-        height: 200,
-        y_mouseover: '%d'
-      },
-      graphData: {}
+      legendLabels: [],
+      values: []
     }
   },
   computed: {
@@ -46,35 +46,330 @@ export default {
     compare () {
       return this.$store.state.compare
     },
+    comparedRepo () {
+      return this.$store.state.comparedRepo
+    },
     rawWeekly () {
       return this.$store.state.rawWeekly
     },
-    chart () {
-      // Set the MetricsGraphics config as much as we can
-      delete this.mgConfig.chart_type
-      this.mgConfig.title = this.title || 'Activity'
-      this.mgConfig.x_accessor = 'date'
-      this.mgConfig.format = this.percentage ? 'percentage' : undefined
-      this.mgConfig.compare = this.compare
-      this.mgConfig.byDate = true
-      this.mgConfig.area = this.rawWeekly
-      this.mgConfig.y_accessor = 'value'
-      this.mgConfig.legend_target = this.$refs.legend
-      this.mgConfig.colors = []
-      this.mgConfig.legend = []
-      this.mgConfig.baselines = []
+    showArea() {
+      return this.$store.state.showArea
+    },
+    showTooltip() {
+      return this.$store.state.showTooltip
+    },
+    spec() {
 
-      this.__download_data = {}
-      this.__download_file = this.mgConfig.title.replace(/ /g, '-').replace('/', 'by').toLowerCase()
+      let config = {
+        "$schema": "https://vega.github.io/schema/vega-lite/v2.json",
+        "title": {
+          "text": this.title,
+          "offset": 15
+        },
+        "width": 420,
+        "height": 200,
+        "config":{
+          "axis":{
+            "grid": false
+          },
+          "legend": {
 
-      // Hide the old chart
-      if (this.$refs.chart) {
-        this.$refs.chart.className = 'linechart loader'
-        window.$(this.$refs.holder).find('.hideme').addClass('invis')
-        window.$(this.$refs.holder).find('.showme').removeClass('invis')
+            "offset": 0,
+            "titleFontSize": 0,
+            "titlePadding": 10
+
+          }
+        },
+        "layer": [],
+        "padding": {
+          "top": 20,
+          "left": 0,
+          "right": 30,
+          "bottom": 55
+        }
       }
-      this.mgConfig.target = document.createElement('div')
 
+      //cannot have duplicate selection, so keep track if it has already been added
+      let selectionAdded = false
+
+      let getStandardLine = function (key) {
+        let color = "FF3647"
+        if (key != "value"){
+          color = "4736FF"
+        }
+        return {
+            "encoding": {
+              "x": {
+                "field": "date",
+                "type": "temporal"
+              },
+              "y": {
+                "field": key,
+                "type": "quantitative"
+              },
+              "color": {
+                "value": color
+              }
+            },
+            "mark": {
+              "type": "line",
+              "interpolate": "basis",
+              "clip": true
+            },
+
+          }
+      }
+
+      let getStandardPoint = function (key) {
+        let selection = {
+              "tooltip": {
+                "type": "single",
+                "nearest": true,
+                "on": "mouseover",
+                "encodings": [
+                  "x"
+                ],
+                "empty": "none"
+              }
+            }
+        if (selectionAdded) {
+          selection = null
+        }
+        selectionAdded = true
+        return {
+            "encoding": {
+              "x": {
+                "field": "date",
+                "type": "temporal",
+                "axis": {
+                  "title": null
+                }
+              },
+              "y": {
+                "field": key,
+                "type": "quantitative",
+                "axis": {
+                  "title": null
+                }
+              },
+              "color": {
+                "field": "name",
+                "type": "nominal",
+                "scale": { "range": ['#4736FF','#FF3647' ] }
+              },
+              "opacity": {
+                "condition": {
+                  "selection": "tooltip",
+                  "value": 1
+                },
+                "value": 0
+              }
+            },
+            "mark": {
+              "type": "point"
+            },
+            "selection": selection
+          }
+      }
+
+    let getArea = function (key) {
+        return {
+              "mark": {
+                "type": "area",
+                "interpolate": "basis",
+                "clip": true
+              },
+              "encoding": {
+                "x": {
+                  "field": "date",
+                  "type": "temporal",
+                  "timeUnit": "year"
+                },
+                "y": {
+                  "aggregate": "ci1",
+                  "field": key,
+                  "type": "quantitative"
+
+                },
+                "y2": {
+                  "aggregate": "ci0",
+                  "field": key,
+                  "type": "quantitative"
+                },
+                "color": {
+                  "value": "green"
+                },
+                "opacity": {"value": 0.2}
+              }
+            }
+          }
+
+    let rule =
+
+      {
+        "transform": [
+        {
+          "filter": {
+            "selection": "tooltip"
+          }
+        }
+      ],
+      "mark": "rule",
+      "encoding": {
+        "x": {
+              "type": "temporal",
+              "field": "date"
+            },
+            "color": {
+              "field": "name",
+              "type": "nominal",
+              "scale": { "range": [ '#4736FF','#FF3647'] }
+            }
+        }
+      }
+
+      let getValueText = function (key){
+          return {
+
+          "transform": [
+            {
+              "filter": {
+                "selection": "tooltip"
+              }
+            }
+          ],
+          "mark": {
+            "type": "text",
+            "align": "left",
+            "dx": 5,
+            "dy": -5
+          },
+          "encoding": {
+            "text": {
+              "type": "quantitative",
+              "field": key
+            },
+            "x": {
+              "type": "temporal",
+              "field": "date"
+            },
+            "y": {
+              "field": key,
+              "type": "quantitative"
+            },
+            "color": {
+                "value": "green"
+              }
+          }
+        }
+      }
+
+      let getDateText = function (key){
+        return {
+
+          "transform": [
+            {
+              "filter": {
+                "selection": "tooltip"
+              }
+            }
+          ],
+          "mark": {
+            "type": "text",
+            "align": "left",
+            "dx": 5,
+            "dy": -15
+          },
+          "encoding": {
+            "text": {
+              "type": "temporal",
+              "field": "date"
+            },
+            "x": {
+              "type": "temporal",
+              "field": "date"
+            },
+            "y": {
+              "field": key,
+              "type": "quantitative"
+            },
+            "color": {
+                "value": "black"
+              }
+          }
+        }
+      }
+
+      //so we can reference the comparedRepo inside of functions ("buildMetric()" specifically)
+      let comparedRepo = this.comparedRepo
+
+      let buildMetric = function () {
+        //build lines and points for initial repo
+        buildLines("value")
+
+        //build lines and points for compared repo
+        if(comparedRepo) buildLines("comparedValue")
+      }
+
+      let buildLines = function (key) {
+        config.layer.push(getStandardLine(key))
+        config.layer.push(getStandardPoint(key))
+      }
+
+      let buildTooltip = function (key) {
+        config.layer.push(getValueText(key))
+        config.layer.push(getDateText(key))
+      }
+
+      buildMetric()
+
+      //push the area to general spec
+      if(this.showArea) {
+        config.layer.push(getArea("value"))
+        if(this.comparedRepo){
+          config.layer.push(getArea("comparedValue"))
+        }
+      }
+      else {
+        //if user doesn't want area mark, then set layers to og
+        for(var x = 0; x < config.layer.length; x++) {
+          if(config.layer[x] == getArea("value")) {
+            buildMetric()
+          }
+        }
+      }
+
+      //push the tooltip to general spec
+      if(this.showTooltip) {
+        buildTooltip("value")
+        //push parts of layer that use "comparedValue" key if there is a comparedRepo
+        if(this.comparedRepo){
+          buildTooltip("comparedValue")
+          config.layer.push(rule)
+        }
+      }
+      else {
+        //if user doesn't want tooltip mark, then iterate through all marks and pop the tooltip marks
+        for(var x = 0; x < config.layer.length; x++) {
+          if(config.layer[x] == getValueText("value")) {
+            buildMetric()
+          }
+        }
+      }
+
+      //set dates from main control options
+      //can either have first priority being to show data points across whole graph, or following user defined dates as first priority
+      for(var i = 0; i < config.layer.length; i++){
+        config.layer[i].encoding.x["scale"] =
+          {
+            "domain": [{"year": this.earliest.getFullYear(), "month": this.earliest.getMonth(), "date": this.earliest.getDate()},{"year": this.latest.getFullYear(), "month": this.latest.getMonth(), "date": this.latest.getDate()}]
+          }
+      }
+
+      let hideRaw = !this.rawWeekly
+
+      $(this.$el).find('.showme').addClass('invis')
+      $(this.$el).find('.linechart').addClass('loader')
       /*
        * Takes a string like "commits,lines_changed:additions+deletions"
        * and makes it into an array of endpoints:
@@ -97,27 +392,40 @@ export default {
         }
       })
 
-      console.log('endp, f', endpoints, fields)
 
       // Get the repos we need
       let repos = []
       if (this.repo) {
         repos.push(window.AugurRepos[this.repo])
       } // end if (this.$store.repo)
-      if (this.comparedTo) {
-        repos.push(window.AugurRepos[this.comparedTo])
+      if (this.comparedRepo) {
+        repos.push(window.AugurRepos[this.comparedRepo])
       }
+
 
       // Make a batch request for all the data we need
       window.AugurAPI.batchMapped(repos, endpoints).then((data) => {
         // Make it so the user can save the data we are using
         this.__download_data = data
+        this.__download_file = this.title.replace(/ /g, '-').replace('/', 'by').toLowerCase()
         this.$refs.downloadJSON.href = 'data:text/json;charset=utf-8,' + encodeURIComponent(JSON.stringify(this.__download_data))
         this.$refs.downloadJSON.setAttribute('download', this.__download_file + '.json')
 
-        // We usually want to limit dates and convert the key to being metrics-graphics friendly
-        let defaultProcess = (obj, key, field, count) => {
-          let d = AugurStats.convertKey(obj[key], field)
+
+        // We usually want to limit dates and convert the key to being vega-lite friendly
+        let defaultProcess = (obj, key, field, count, compared) => {
+          // let a = null
+          //let b = AugurStats.convertKey(obj[key], field)
+          let d = null
+          if(compared) {
+            d = AugurStats.convertComparedKey(obj[key], field)
+            //d = b.concat(a)
+
+          }
+          else {
+            d = AugurStats.convertKey(obj[key], field)
+          }
+
           d = AugurStats.convertDates(d, this.earliest, this.latest)
           return d
         }
@@ -125,18 +433,19 @@ export default {
         // Normalize the data into [{ date, value },{ date, value }]
         // BuildLines iterates over the fields requested and runs onCreateData on each
         let normalized = []
-        let buildLines = (obj, onCreateData) => {
+        let aggregates = []
+        let buildLines = (obj, onCreateData, compared) => {
           if (!obj) {
             return
           }
           if (!onCreateData) {
             onCreateData = (obj, key, field, count) => {
-              let d = defaultProcess(obj, key, field, count)
               normalized.push(d)
             }
           }
           let count = 0
           for (var key in obj) {
+
             if (obj.hasOwnProperty(key)) {
               if (fields[key]) {
                 fields[key].forEach((field) => {
@@ -157,71 +466,130 @@ export default {
           } // end for in
         } // end normalize function
 
+
         // Build the lines we need
-        if (!this.comparedTo) {
+        let legend = []
+        let values = []
+        let colors = []
+        if (!this.comparedRepo) {
           buildLines(data[this.repo], (obj, key, field, count) => {
             // Build basic chart using rolling averages
-            let d = defaultProcess(obj, key, field, count)
+            let d = defaultProcess(obj, key, field, count, false)
             let rolling = AugurStats.rollingAverage(d, 'value', this.period)
-            console.log('rolling, before+after', d, rolling)
             if (!this.disableRollingAverage) {
               normalized.push(rolling)
-              this.mgConfig.legend.push(field)
-              this.mgConfig.colors.push(window.AUGUR_CHART_STYLE.brightColors[count])
+              aggregates.push(d)
+              legend.push(field)
+              colors.push(window.AUGUR_CHART_STYLE.brightColors[count])
             }
-            if (this.rawWeekly || this.disableRollingAverage) {
-              normalized.push(d)
-              this.mgConfig.legend.push(field)
-              this.mgConfig.colors.push(this.disableRollingAverage ? window.AUGUR_CHART_STYLE.brightColors[count] : window.AUGUR_CHART_STYLE.dullColors[count])
+            if (!hideRaw || this.disableRollingAverage) {
+              normalized.push(rolling)
+              aggregates.push(d)
+              legend.push(field)
+              colors.push(this.disableRollingAverage ? window.AUGUR_CHART_STYLE.brightColors[count] : window.AUGUR_CHART_STYLE.dullColors[count])
             }
-          })
-        } else if (this.compare === 'each' && this.comparedTo) {
+          }, false)
+        } else if (this.compare === 'each' && this.comparedRepo) {
           // Build comparison using z-scores
-          buildLines(data[this.comparedTo], (obj, key, field, count) => {
-            let d = defaultProcess(obj, key, field, count)
-            let rolling = AugurStats.rollingAverage(AugurStats.zscores(d, 'value'), 'value', this.period)
-            normalized.push(rolling)
-            this.mgConfig.legend.push(this.comparedTo + ' ' + field)
-            this.mgConfig.colors.push(window.AUGUR_CHART_STYLE.brightColors[count])
-          })
           buildLines(data[this.repo], (obj, key, field, count) => {
-            let d = defaultProcess(obj, key, field, count)
-            let rolling = AugurStats.rollingAverage(AugurStats.zscores(d, 'value'), 'value', this.period)
+            let d = defaultProcess(obj, key, field, count, false)
+            let rolling = AugurStats.rollingAverage(d, 'value', this.period)
+
+            //let rolling = AugurStats.rollingAverage(AugurStats.zscores(d, 'value'), 'value', this.period)
             normalized.push(rolling)
-            this.mgConfig.legend.push(this.repo + ' ' + field)
-            this.mgConfig.colors.push(window.AUGUR_CHART_STYLE.dullColors[count])
-          })
-        } else if (this.comparedTo) {
+            aggregates.push(d)
+            legend.push(this.repo + ' ' + field)
+            colors.push(window.AUGUR_CHART_STYLE.dullColors[count])
+          }, false)
+          buildLines(data[this.comparedRepo], (obj, key, field, count) => {
+            let d = defaultProcess(obj, key, field, count, true)
+            let rolling = AugurStats.rollingAverage(d, 'comparedValue', this.period)
+            //let rolling = AugurStats.rollingAverage(AugurStats.zscores(d, 'comparedValue'), 'comparedValue', this.period)
+            normalized.push(rolling)
+            aggregates.push(d)
+            legend.push(this.comparedRepo + ' ' + field)
+            colors.push(window.AUGUR_CHART_STYLE.brightColors[count])
+          }, true)
+        } else if (this.comparedRepo) {
           // Build chart compared to baseline
-          this.mgConfig.baselines = [{value: 1, label: this.repo}]
-          buildLines(data[this.comparedTo], (obj, key, field, count) => {
+          //this.mgConfig.baselines = [{value: 1, label: this.repo}]
+          buildLines(data[this.comparedRepo], (obj, key, field, count) => {
             normalized.push(AugurStats.makeRelative(obj[key], data[this.repo][key], field, {
               earliest: this.earliest,
               latest: this.latest,
               byDate: true,
               period: this.period
             }))
-            this.mgConfig.legend.push(this.comparedTo + ' ' + field)
-            this.mgConfig.colors.push(window.AUGUR_CHART_STYLE.brightColors[count])
-          })
+            legend.push(this.comparedRepo + ' ' + field)
+            colors.push(window.AUGUR_CHART_STYLE.brightColors[count])
+          }, true)
         }
 
         if (normalized.length == 0) {
-          this.mgConfig.missing_text = 'Data empty'
           this.renderError()
         } else {
-          this.mgConfig.data = normalized
-          this.mgConfig.legend_target = this.$refs.legend
+          //shared.baseData = data.map((e) => { e.repo = this.repo.toString(); return e })
+
+
+          if(hideRaw) {
+            for(var i = 0; i < legend.length; i++){
+              normalized[i].forEach(d => {
+                d.name = legend[i]
+                d.color = colors[i]
+                values.push(d);
+              })
+            }
+          }
+          else {
+            for(var i = 0; i < legend.length; i++){
+              aggregates[i].forEach(d => {
+                d.name = "raw " + legend[i]
+                d.color = colors[i]
+                values.push(d)
+              })
+            }
+          }
+
+          //if(this.comparedRepo){this.legendLabels = legend}
+          //else{for (var label in legend){this.legendLabels.push(label[0])}}
+          this.legendLabels = legend
+          this.values = values
+
+          //function getMaxY(arr){
+            // var temp = normalized[0]
+            // console.log(temp[0].value)
+            // var output = [];
+            // for (var i=0; i < temp.length ; ++i)
+            //     if(temp[i]) output.push(temp[i][value]);
+            // console.log("output" + output);
+            // var result = objArray.map(temp => temp.value);
+            // console.log(result)
+            // console.log(normalized[0])
+            //return result[0][0].reduce((max, b) => Math.max(max, b.value), data[0].value);
+          //}
+          //console.log(getMaxY())
+          //console.log("LOOK " + normalized[1][0].value)
+          config.config.legend.offset = -(String(this.legendLabels[0]).length * 6.5) - 20
+          //console.log("range " + normalized[0][0].value)
+          //if(this.values[this.values.length - 1].value > Math.max.apply(Math, normalized) * .6) {
+            //console.log("IT IS HAPPENING HAHAH")
+            //config.config.legend.titlePadding = 175
+          //}
+
+          $(this.$el).find('.showme, .hidefirst').removeClass('invis')
+          $(this.$el).find('.linechart').removeClass('loader')
+          //this.mgConfig.legend_target = this.$refs.legend
           this.renderChart()
         }
-        
+
       }, () => {
-        this.mgConfig.missing_text = 'Data is missing or unavaliable'
+        //this.mgConfig.missing_text = 'Data is missing or unavaliable'
         this.renderError()
       }) // end batch request
+      //return '<div class="loader deleteme">' + this.title + '...</div>'
+      return config
+    },
 
-      return '<div class="loader deleteme">' + this.title + '...</div>'
-    } // end chart()
   }, // end computed
   methods: {
     downloadSVG (e) {
@@ -239,25 +607,25 @@ export default {
       window.$(this.$refs.holder).find('.hideme').removeClass('invis')
       window.$(this.$refs.holder).find('.showme').removeClass('invis')
       window.$(this.$refs.holder).find('.deleteme').remove()
-      window.$(this.mgConfig.target).hover((onEnterEvent) => {
-        window.$(this.$refs.legend).hide()
-      }, (onLeaveEvent) => {
-        window.$(this.$refs.legend).show()
-      })
+      // window.$(this.mgConfig.target).hover((onEnterEvent) => {
+      //   window.$(this.$refs.legend).hide()
+      // }, (onLeaveEvent) => {
+      //   window.$(this.$refs.legend).show()
+      // })
       this.$refs.chartholder.innerHTML = ''
       this.$refs.chartholder.appendChild(this.mgConfig.target)
-      this.mgConfig.target.className = 'deleteme'
-      window.MG.data_graphic(this.mgConfig)
+      //this.mgConfig.target.className = 'deleteme'
+      //window.MG.data_graphic(this.mgConfig)
     },
     renderError () {
       this.$refs.chart.className = 'linechart intro'
       window.$(this.$refs.holder).find('.deleteme').remove()
       this.$refs.chartholder.innerHTML = ''
       this.$refs.chartholder.appendChild(this.mgConfig.target)
-      this.mgConfig.target.className = 'deleteme'
-      this.mgConfig.chart_type = 'missing-data'
-      window.MG.data_graphic(this.mgConfig)
+      //this.mgConfig.target.className = 'deleteme'
+      //this.mgConfig.chart_type = 'missing-data'
+      //window.MG.data_graphic(this.mgConfig)
     }
   }// end methods
-} // end export default {}
+}
 </script>
