+++ frontend/app/styles/ghdata.styl
@@ -16,12 +16,12 @@ header
 header .search
   width: 100%
   margin-top: 25px
-  
+
 #logo
   display: inline-block
   margin-top: 0px
   max-width: 250px
-  
+
 .content
   margin: auto
   width: 95%
@@ -31,7 +31,7 @@ header .search
   color: #000
   background: #eaeaea
   border-bottom: 1px solid #eaeaea
-  
+
 
 section#landing h1
   font-size: 25px
@@ -85,19 +85,19 @@ section.unmaterialized
     padding-bottom: 100px
     position: relative
     z-index: 9
-  
+
   section
     z-index: 10
-  
+
   #landing
     width:100%
-    
+
   .fixed-container
     z-index: 0
     position: absolute
     top: 245px
     left: calc(50% - (540px))
-  
+
   .fixed
     position: fixed !important
     top: 25px
@@ -128,7 +128,7 @@ cite.metric
 cite.metric a
   color: #aaa
   text-decoration: underline;
-  
+
 .graph-download
   background: 0
   color: #ccc
@@ -138,7 +138,7 @@ cite.metric a
   float: right
   margin: -7px 7px 0 0 !important;
   padding: 3px !important;
-  
+
 .graph-download:hover
   background: #eee;
   color: #000
@@ -173,7 +173,7 @@ cite.metric a
 .mg-line-legend
   display: none
 
-.legend 
+.legend
   position: relative
   margin: 0
   padding: 0
@@ -201,8 +201,8 @@ cite.metric a
     filter: grayscale(0%)
   100%
     opacity: 0.2
-    filter: grayscale(100%) 
-    
+    filter: grayscale(100%)
+
 @keyframes loader
   0%
     opacity: 0.2
@@ -216,7 +216,7 @@ cite.metric a
 
 .intro
   animation: intro 2s
-    
+
 @keyframes intro
   0%
     opacity: 0.2
@@ -237,7 +237,7 @@ cite.metric a
     display: none
 
 .hidden, .hidden > *, .hidden > small
-  display: none !important 
+  display: none !important
 
 .invis, .invis > *, .invis > small
   z-index: -1;
